+++ tf/advanced/links.py
@@ -251,7 +251,7 @@ def header(app, allMeta=False):
                 colofon,
                 commonHeader,
                 '<img class="hdlogo" src="/data/static/logo.png"/>',
-                '<img class="hdlogo" src="/server/static/icon.png"/>',
+                '<img class="hdlogo" src="/browser/static/icon.png"/>',
             )
     else:
         tfLine = "\n\t".join(x for x in (tfLink, appLink, tfsLink) if x)